refactor: create a custom hook to handle the timer and make timer component more readable
