+++ src/lib/context/ColorContext.tsx
@@ -3,6 +3,8 @@ import { type ReactNode, createContext, useState } from 'react'
 export interface ColorContextProps {
   color: string
   setColor: (color: string) => void
+  textColor: string
+  setTextColor: (colorClassName: string) => void
 }
 
 interface Props {
@@ -13,10 +15,13 @@ const ColorContext = createContext<ColorContextProps | null >(null)
 
 export const ColorProvider = ({ children }: Props) => {
   const [color, setColor] = useState('#F87070')
+  const [textColor, setTextColor] = useState('hover:text-[#F87070]')
 
   const value = {
     color,
-    setColor
+    setColor,
+    textColor,
+    setTextColor
   }
 
   return (
